+++ src/includes/classes/Core/Utils/Sha256.php
@@ -27,8 +27,11 @@ class Sha256 extends Classes\Core\Base\Core
      *
      * @return string SHA-256 signature string (64 bytes).
      */
-    public function keyedHash(string $string, string $key): string
+    public function keyedHash(string $string, string $key = ''): string
     {
+        if (!$key && !($key = $this->App->Config->©encryption['©key'])) {
+            throw $this->c::issue('Missing encryption key.');
+        }
         return hash_hmac('sha256', $string, $key);
     }
 }